# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cdd-ctl",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/rpc": {
      "post": {
        "tags": [
          "rpc"
        ],
        "summary": "The core JSON-RPC request handler",
        "operationId": "rpc_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RpcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RpcResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RpcError": {
        "type": "object",
        "description": "A JSON-RPC 2.0 Error object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "RpcRequest": {
        "type": "object",
        "description": "A JSON-RPC 2.0 Request payload",
        "required": [
          "jsonrpc",
          "method"
        ],
        "properties": {
          "id": {
            "type": [
              "object",
              "null"
            ],
            "description": "Request ID (optional)"
          },
          "jsonrpc": {
            "type": "string",
            "description": "JSON-RPC version (must be \"2.0\")",
            "example": "2.0"
          },
          "method": {
            "type": "string",
            "description": "The method to invoke",
            "example": "version"
          },
          "params": {
            "type": [
              "object",
              "null"
            ],
            "description": "Parameters for the method (optional)"
          }
        }
      },
      "RpcResponse": {
        "type": "object",
        "description": "A JSON-RPC 2.0 Response payload",
        "required": [
          "jsonrpc"
        ],
        "properties": {
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RpcError",
                "description": "Error payload"
              }
            ]
          },
          "id": {
            "type": [
              "object",
              "null"
            ],
            "description": "Correlated request ID"
          },
          "jsonrpc": {
            "type": "string",
            "description": "JSON-RPC version (always \"2.0\")",
            "example": "2.0"
          },
          "result": {
            "type": [
              "object",
              "null"
            ],
            "description": "Successful result payload"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}